refactor: getRssItem function
