+++ lib/v2/lifeweek/channel.js
@@ -1,11 +1,10 @@
 const got = require('@/utils/got');
-const { parseDate } = require('@/utils/parse-date');
-const timezone = require('@/utils/timezone');
+const getRssItem = require('./utils');
 
 const rootApiUrl = 'https://www.lifeweek.com.cn/api/userWebFollow/getFollowTagContentList?type=3&sort=2&tagId';
 const rootUrl = 'https://www.lifeweek.com.cn/column';
 const articleRootUrl = 'https://www.lifeweek.com.cn/article';
-const articleApiRootUrl = 'https://www.lifeweek.com.cn/api/article';
+
 module.exports = async (ctx) => {
     const channel = ctx.params.id;
     const url = `${rootApiUrl}=${channel}`;
@@ -14,17 +13,7 @@ module.exports = async (ctx) => {
     const items = await Promise.all(
         result.map((item) => {
             const articleLink = `${articleRootUrl}/${item.id}`;
-            const articleApiLink = `${articleApiRootUrl}/${item.id}`;
-            return ctx.cache.tryGet(articleApiLink, async () => {
-                const { data } = await got(articleApiLink);
-                const time = timezone(parseDate(item.pubTime), +8);
-                return {
-                    title: item.title,
-                    description: data.model.content,
-                    link: articleLink,
-                    pubDate: time,
-                };
-            });
+            return ctx.cache.tryGet(articleLink, () => getRssItem(item, articleLink));
         })
     );
