+++ lib/v2/lifeweek/tag.js
@@ -1,11 +1,9 @@
 const got = require('@/utils/got');
-const { parseDate } = require('@/utils/parse-date');
-const timezone = require('@/utils/timezone');
-
+const getRssItem = require('./utils');
 const rootApiUrl = 'https://www.lifeweek.com.cn/api/userWebFollow/getFollowTagContentList?type=4&sort=2&tagId';
 const rootUrl = 'https://www.lifeweek.com.cn/articleList';
 const articleRootUrl = 'https://www.lifeweek.com.cn/article';
-const articleApiRootUrl = 'https://www.lifeweek.com.cn/api/article';
+
 module.exports = async (ctx) => {
     const tag = ctx.params.id;
     const url = `${rootApiUrl}=${tag}`;
@@ -14,19 +12,7 @@ module.exports = async (ctx) => {
     const items = await Promise.all(
         result.map((item) => {
             const articleLink = `${articleRootUrl}/${item.id}`;
-            const articleApiLink = `${articleApiRootUrl}/${item.id}`;
-            return ctx.cache.tryGet(articleApiLink, async () => {
-                const { data } = await got(articleApiLink, {
-                    responseType: 'buffer',
-                });
-                const time = timezone(parseDate(item.pubTime), +8);
-                return {
-                    title: item.title,
-                    description: data.model.content,
-                    link: articleLink,
-                    pubDate: time,
-                };
-            });
+            return ctx.cache.tryGet(articleLink, () => getRssItem(item, articleLink));
         })
     );
